Change mutant operater location
