--- core/src/test/scala/stryker4s/mutants/findmutants/MutantCreatorTest.scala
@@ -1,19 +1,0 @@
-package stryker4s.mutants.findmutants
-
-import stryker4s.Stryker4sSuite
-import stryker4s.extensions.ImplicitMutationConversion.mutationToTree
-import stryker4s.extensions.mutationtypes.{EqualTo, GreaterThan, GreaterThanEqualTo, LesserThan}
-
-class MutantCreatorTest extends Stryker4sSuite {
-
-  case class Sut() extends MutantCreator
-
-  describe("mutant id") {
-    it("should register multiple mutants from a FoundMutant with multiple mutations") {
-      val sut = Sut()
-      val mutants = sut.TermExtensions(GreaterThan) ~~> (LesserThan, GreaterThanEqualTo, EqualTo)
-
-      mutants.map(mutant => mutant.id) should contain theSameElementsAs List(0, 1, 2)
-    }
-  }
-}
